mcp_toolbox/pyproject.toml
[tool.poetry]
name = "mcp-toolbox"
version = "0.1.0"
description = "MCP toolbox server for Cloud SQL and Redis integration with Google ADK agents"
authors = ["Mridula Vinod <mridulait67@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    {include = "mcp_tools"}
]

[tool.poetry.dependencies]
python = "^3.10"
google-adk = { version = "==1.14.1", extras = ["eval"] }
google-cloud-aiplatform = {extras = ["adk", "agent-engines"], version = "^1.88.0"}
google-cloud-sql-connector = "^1.0.0"
google-cloud-redis = "^2.15.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.0"
sqlalchemy = "^2.0.23"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
asyncio-redis = "^0.16.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"